fixed compilation errors

+++ Core/Subscale/DenseUnitGenerator/DenseUnit.cpp
@@ -16,7 +16,7 @@ uint64_t DenseUnit::getSignature() const
     return signature_;
 }
 
-uint32_t DenseUnit::getDimension() const
+uint64_t DenseUnit::getDimension() const
 {
     return dimension_;
 }
